Turn font-variant property into a shorthand
Font-variant in the old CSS 2.1 specificiation was a primitive value,
either normal, inherit, or small-caps.

In the latest CSS specification font-variant is a shorthand for the
font-variant-* subproperties, when used in a style declaration. [2]

Also, font-variant is a font-face descriptor when used in a font-face
style rule, where values are not expanded into subproperties or
sub-descriptors, but uses the same values as for the subproperties [3].

In this CL, parsing, computed style value mapping, popup menu
implementation, tests and other parts of the code are updated to reflect
interpreting font-variant as a shorthand.

Also, Variant is removed from FontTraits, which reflects the CSS
specification more correctly, as the variant parameter is not part of
font matching.

[1] https://www.w3.org/TR/2011/REC-CSS2-20110607/fonts.html#small-caps
[2] https://drafts.csswg.org/css-fonts/#font-variant-prop
[3] https://drafts.csswg.org/css-fonts/#font-rend-desc

BUG=587094,516673,585746
TEST=fast/text/font-features/font-variant-shorthand.html

Review-Url: https://codereview.chromium.org/1900603002
Cr-Original-Commit-Position: refs/heads/master@{#391245}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f5c3fca69f7b1fc2e2ae6c51797e8714c36acaf9