Cache hb_font_t and the user data container instead of hb_face_t
Previously, we did cache the hb_face_t but not the hb_font_t, and
instead generated an hb_font_t for each HarfBuzzFace object,
outside of the cache. In the prepareHarfBuzzFontData() function,
we also reset the font functions on the newly created
hb_font_t. So in the HarfBuzzFace constructor and in each created
HarfBuzzFace object, we would create hb_font_t and reset it,
instead of reusing this from the cache, too.

This lead to HarfBuzz reloading font tables for each
FontPlatformData instance, instead of using the
per-unique-typeface caching mechanism.

This should address the memory regression we have been observing
on Android.

BUG=577306
R=behdad,eae

Review-Url: https://codereview.chromium.org/1950183004
Cr-Original-Commit-Position: refs/heads/master@{#391835}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3fa6af33bc74e5c332f8ee63efc40536ff027538