Re-enable complex text path on android
Trying to re-enable the complex text path on Android now that hb_font_t
is cached with the user data container instead of the hb_face_t object.

This should result in lower memory usage for the complex text code path,
we'll watch the memory bots though and revert as needed if it regresses.

BUG=577306
R=drott@chromium.org

Review-Url: https://codereview.chromium.org/1949053002
Cr-Original-Commit-Position: refs/heads/master@{#392010}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9af72a88ea5b39c2c36c99f8b091c5feae82ef7d