Keep PaymentRequest alive when its promises are outstanding.
The PaymentRequest is responsible for resolving the show and complete
promises, so it needs to be kept alive as long as either of these
promises haven't resolved yet. This makes sure of that.

Also reverts the earlier change that "stopped" outstanding promises
when the context is destroyed, as that doesn't actually do anything.

BUG=606577

Review-Url: https://codereview.chromium.org/1954973005
Cr-Original-Commit-Position: refs/heads/master@{#392157}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2fc9122808b627241860773ec97c8d9120df7ee