[css-grid] Store auto repeat type for both rows and columns
We're storing in style the the auto-repeat type (auto-fill or auto-fit) but
using just one slot. That was wrong because rows and columns could both have
auto-repeat tracks with different types (still just one auto-repeat per
axis).

This was not causing any issue because we're not using it yet until auto-fit
is not implemented. That's the reason why we cannot create a test for it. In
any case the future implementation of auto-fit won't ever work without this.

BUG=589460

Review-Url: https://codereview.chromium.org/1952923004
Cr-Original-Commit-Position: refs/heads/master@{#392321}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a2a9aa6891dc3fb5e67a4379e3a132a6a17ef937