Implement offscreenCanvas.getContext('webgl') on a worker thread
Right now, offscreenCanvas.getContext('webgl') works on the main thread
only because it requires GpuChannelHost. In order to make it work on a
worker thread, we let the worker thread issue a waitable event, and
postTask() to main thread, ask main thread to create the contextProvider,
and signal back to the worker thread.

In this CL, we also Make BindToCurrentThread() optional in
createOffscreenGraphicsContext3DProvider.
Obtain a contextProvider on a worker thread involves these steps:
1. postTask to the main thread.
2. main thread calls createOffscreenGraphicsContext3DProvider, which
   implicitly calls BindToCurrentThread().
3. main thread signal that contextProvider is ready.
4. worker thread calls contextProvider->DetachFromThread(), and then
   contextProvider->BindToCurrentThread().

We can see that BindToCurrentThread() gets called twice. If we make
the BindToCurrentThread() call optional in createOffscreenGraphicsContext3DProvider,
then the worker thread will be the first one calling BindToCurrentThread(),
and there will be no need to call DetachFromThread().

Besides that, we made the the PersistentHeapHashSet and PersistentHeapHashMap
in the WebGLRenderingContextBase to be per-thread. The reason is
that we do not want the contexts created on the main thread and on
worker threads to share the same hashset or hashmap.

BUG=602391
CQ_INCLUDE_TRYBOTS=tryserver.chromium.win:win_optional_gpu_tests_rel;tryserver.chromium.mac:mac_optional_gpu_tests_rel

Review-Url: https://codereview.chromium.org/1914233006
Cr-Original-Commit-Position: refs/heads/master@{#392425}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2ec178e3f4f22a0806b5a13aff43a8d570e6c44c