Allow expensive task blocking if there is pending iframe navigation
The motivation for this patch is to gain the benefits of the expensive
task blocking intervention even if an iframe is navigating.  This
should help improve touch responsiveness on ad heavy pages.

We still prevent expensive task blocking when there is a pending main
frame navigation.

BUG=600202

Review-Url: https://codereview.chromium.org/1962053002
Cr-Original-Commit-Position: refs/heads/master@{#392583}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb4be51059b55cd04c329d2d296ebcc7e808b215