Address more TODO(GYP) in the GN build.
Addresses another batch of TODO(GYP) annotations in GN files.

Removes all references to msvs_shard which is not necesasary in the GN build.

A number of these were related to things to support the nacl64 build. In GN we just do cross-compiles of the fill targets. This seems to be working OK. Probably the sizes can be brought down a bit by optimizing this and producing nacl-specific targets with only the bare necessities, but it seems the marginal benefit doesn't outweigh the build complexity.

TBR=thestig@chromium.org for linux/build (comment change)

Review-Url: https://codereview.chromium.org/1957683002
Cr-Original-Commit-Position: refs/heads/master@{#392660}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57399c53d558cc84c865ad48a211fe9e92c2ccf9