Add support for origin trials in workers.
This adds support for origin trials in workers (dedicated, shared
and service workers). It does so by adding a list of tokens to
WorkerThreadStartupData. That list is filled using the response headers
for shared and service workers, and filled with the documents tokens
for dedicated workers.

BUG=582063
BUG=582195
BUG=582197

Review-Url: https://codereview.chromium.org/1828063002
Cr-Original-Commit-Position: refs/heads/master@{#392661}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de1993ee636b9c8a2ca0f1a817ff3d3ca5c88a6d