Update third_party/binutils to 2.26, add ICF fix
The new binutils 2.26 release incorporates two out of the three
fixes that we were applying to the 2.25 version.  There is also
an upstream fix, that's not in 2.26, for a bug that made the
linker crash when using --icf for x86-32, which we now apply
locally so it should become safe to use gold with --icf for x86.

BUG=610908
R=thakis@chromium.org

Review-Url: https://codereview.chromium.org/1950343004
Cr-Original-Commit-Position: refs/heads/master@{#392828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a1fcea64a6130271e68cddb71fb71f71080c3e2