RTCCertificate serialization - persistent storage in IndexedDB.
This is part of larger work that will allow persistent storage of
RTCCertificate in IndexedDB. The big picture can be seen here:
https://codereview.chromium.org/1949033002/

This CL implements part 2 and 3c:

2. RTCCertificate serialization (structured clone algorithm)
   a) ScriptValueSerializerForModules and friends are updated with write and
      read functions for RTCCertificate (impl using 1.).

3. Unittests making sure RTCCertificate can be saved and loaded from IndexedDB
   c) WebRtcBrowserTest tests added which clone certificates and sets up calls
      using the clones.

This is the final piece of the puzzle, pt 1 and 3a-b have already landed.
Landing this turns the feature on.

BUG=581354

Review-Url: https://codereview.chromium.org/1959333002
Cr-Original-Commit-Position: refs/heads/master@{#392893}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8730debe7de67ffc97aed74a5cd7f0cd799cb60