Avoid object slicing in WebCredential::create
WebCredential::create currently returns a WebCredential. It tries to create one
of its subclasses, WebPasswordCredential or WebFederatedCredential,
respectively, but those object get sliced to their common base upon returning,
because a function returning A cannot return B even if B is a subclass of A.

To fix that, this CL changes create() to return a pointer to a WebCredential.
This allows creating subclasses of it while keeping the return type as
declared.

R=mkwst@chromium.org
BUG=610646

Review-Url: https://codereview.chromium.org/1967693003
Cr-Original-Commit-Position: refs/heads/master@{#392959}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9818ab57e995473e2d41d5bb40b6a9abd2682151