Remove unused imports in webkitpy with using autoflake.
This is a slightly more conservative re-do of
https://codereview.chromium.org/1847613002.

For this CL, I ran `autoflake -r -i webkitpy` and then undid any changes
to anything in thirdparty or __init__.py files (which can be dealt
with separately).

After https://codereview.chromium.org/1847613002, the
problem that happened was due to removing the imports in
third_party/WebKit/Tools/Scripts/webkitpy/tool/commands/__init__.py,
which made it so that multicommandtool was unable to find the subclasses
of Command, which broke webkit-patch (which is used by the auto-rebaseline
bot).

This CL doesn't remove anything from webkitpy/tool/commands/__init__.py
or webkitpy/tool/main.py.

BUG=598897

Review-Url: https://codereview.chromium.org/1950213002
Cr-Original-Commit-Position: refs/heads/master@{#393077}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 33305ad2a855ad0269e5b33bc1948e8c7d8e7a76