Change the return type of WTF::bind() to std::unique_ptr.
Yes, this touches a lot of files. I think they are the transitive
closure of compile errors originating from the return type of
WTF::bind().

This is kind of a testbed or an early experimentation to figure
out the remaining issues we need to resolve before we reach out
to blink-dev for help.

BUG=582354

Review-Url: https://codereview.chromium.org/1968613004
Cr-Original-Commit-Position: refs/heads/master@{#393218}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c454ebb21f38ab3de1202b1cade1508ab11cd426