Fix crash derefing null RootScroller
Not sure how we end up with a null RootScroller here since we're in
the main frame and its local (since we're in Document). In any case,
I've rearranged the ownership of RootScroller in
https://codereview.chromium.org/1970763002/ which should simplify the
lifetime issues but I won't be able to land that soon so this CL is
meant as a stop gap until that does land.

BUG=611297

Review-Url: https://codereview.chromium.org/1969343003
Cr-Original-Commit-Position: refs/heads/master@{#393258}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91e80d28ca509e48d697603fe803d63e4b87921d