Support includePartialGlyphs=false in Font::offsetForPositionForComplexText
This patch supports includePartialGlyphs=false in
Font::offsetForPositionForComplexText(). This flag was supported in
offsetForPositionForSimpleText(), but was assumed to be always true in
offsetForPositionForComplexText().

Over 70 existing tests fail by the DCHECK added in this CL without this
fix when the hack in rewindToMidWordBreak() was removed.

BUG=610791

Review-Url: https://codereview.chromium.org/1966703002
Cr-Original-Commit-Position: refs/heads/master@{#393300}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ad66862296c227c1b74a15f96632a67d941b6643