Split Command and HelpCommand into separate files.
This is a refactoring CL.

Reason: The base Command class can logically be considered and tested separately from MultiCommandTool.

This is also somewhat related to clearing out the imports from commands/__init__.py -- in order to not do imports there, we could explicitly import and enumerate the Command classes to include, either in webkit_patch.py or multicommandtool.py, and so individual Command classes shouldn't depend on the multicommandtool module (because that leads to circular imports).

Review-Url: https://codereview.chromium.org/1971593002
Cr-Original-Commit-Position: refs/heads/master@{#393777}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5bd243eea2ee9e7dc3a294e201defc63d55bfd0a