Remove StepSequence and AbstractSequencedCommand.
Reason: The only command that subclasses AbstractSequencedCommand now is PrettyDiff, and PrettyDiff is no longer actually a sequence of steps (it's only one step), so the code related to steps and step sequences can be removed.

This was suggested in http://crrev.com/1971593002.

Review-Url: https://codereview.chromium.org/1985553002
Cr-Original-Commit-Position: refs/heads/master@{#394067}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e83219cd5e9db35e916cda86cf0e085316821cb