Inflate dirty rect by 1 pixel on a high DPI display
On a high dpi display, putImageData leaves visible artifacts around the
affected area's border.

This CL fixes it in the way that when there is antialias involved, the
dirty rect is inflated by 1 pixel.

Locally we can see this bug goes away with this fix on a retina display.
Note that this bug doesn't repro with our layout test. So a manual
test is added to prevent regression.

BUG=609820

Review-Url: https://codereview.chromium.org/1983863002
Cr-Original-Commit-Position: refs/heads/master@{#394286}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0baf3521c90b2efda95e33c5ed26cc2c51423381