MediaStream Image Capture (7): add minimal PhotoCapabilities
This CL continues implementing ImageCapture, concretely:

- Adds the field |photoCapabilities|, of type
PhotoCapabilities, for which the .idl, .h are also
added. For the purpose of this CL, only one field,
 |zoom|, is added.

- Said |zoom| is of type MediaSettingsRange, and an
.idl and .h are added for it.

- ImageCapture.cpp launches a capabilities request on
construction, and adds a template to convert between
mojo::Callback and wtf execution.

- the .mojom file is updated, as is the content/
implementation (returns nothing) and the mock used
for LayoutTests.
- A new LayoutTest is added based on said mock.

BUG=518807

Review-Url: https://codereview.chromium.org/1971233003
Cr-Original-Commit-Position: refs/heads/master@{#394326}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ebfb2e861653cec5d2e371efa7abe96feacbc4b